[refactor] 폴더 구조 변경 및 공백 제거
