+++ BE/src/main/java/cuketmon/battle/constant/BattleResult.java
@@ -1,9 +1,9 @@
-package cuketmon.battle;
+package cuketmon.battle.constant;
 
 public enum BattleResult {
 
     TRAINER1_WIN,
     TRAINER2_WIN,
     DRAW
-    
+
 }
